# docs/swagger.yaml
definitions:
  model.CategoryCreate:
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  model.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.CategoryUpdate:
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  utility.ResponseError:
    properties:
      error:
        type: string
    type: object
  utility.ResponseSuccess:
    properties:
      data: {}
    type: object
info:
  contact: {}
paths:
  /api/category:
    get:
      description: Retrieve a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CategoryResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: List all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new category with the provided details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Create a new category
      tags:
      - Category
  /api/category/{id}:
    delete:
      description: Delete a specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Delete a category
      tags:
      - Category
    get:
      description: Retrieve details of a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Get category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update the details of a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Update a category
      tags:
      - Category
...
swagger: "2.0"